Inner edge into modify orbits direct/forces
