+++ src/modify_orbits_forces.c
@@ -75,13 +75,17 @@
 #include "rebxtools.h"
 
 static struct reb_vec3d rebx_calculate_modify_orbits_forces(struct reb_simulation* const sim, struct rebx_force* const force, struct reb_particle* p, struct reb_particle* source){
-    double tau_a = INFINITY;
+    double invtau_a = 0.0;
     double tau_e = INFINITY;
     double tau_inc = INFINITY;
     
     const double* const tau_a_ptr = rebx_get_param(sim->extras, p->ap, "tau_a");
     const double* const tau_e_ptr = rebx_get_param(sim->extras, p->ap, "tau_e");
     const double* const tau_inc_ptr = rebx_get_param(sim->extras, p->ap, "tau_inc");
+
+    //Implement the planet trap
+    const double* const dedge = rebx_get_param(sim->extras, force->ap, "inner_disk_edge_position");
+    const double* const hedge = rebx_get_param(sim->extras, force->ap, "disk_edge_width");
 
     const double dvx = p->vx - source->vx;
     const double dvy = p->vy - source->vy;
@@ -92,7 +96,13 @@ static struct reb_vec3d rebx_calculate_modify_orbits_forces(struct reb_simulatio
     const double r2 = dx*dx + dy*dy + dz*dz;
     
     if(tau_a_ptr != NULL){
-        tau_a = *tau_a_ptr;
+        invtau_a = 1.0/(*tau_a_ptr);
+        if ((dedge!=NULL)&(hedge!=NULL)){
+            int err=0;
+            struct reb_orbit o = reb_tools_particle_to_orbit_err(sim->G, *p, *source, &err);
+            const double a0 = o.a;
+            invtau_a *= rebx_calculate_planet_trap(a0, *dedge, *hedge);
+        }
     }
     if(tau_e_ptr != NULL){
         tau_e = *tau_e_ptr;
@@ -103,9 +113,9 @@ static struct reb_vec3d rebx_calculate_modify_orbits_forces(struct reb_simulatio
     
     struct reb_vec3d a = {0};
 
-    a.x =  dvx/(2.*tau_a);
-    a.y =  dvy/(2.*tau_a);
-    a.z =  dvz/(2.*tau_a);
+    a.x =  dvx*invtau_a/(2.);
+    a.y =  dvy*invtau_a/(2.);
+    a.z =  dvz*invtau_a/(2.);
 
     if (tau_e < INFINITY || tau_inc < INFINITY){
         const double vdotr = dx*dvx + dy*dvy + dz*dvz;
